+++ app/Http/Controllers/Admin/LearningPlanController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\CompanyEmployee;
 use App\Models\MyLearningPlan;
+use App\Models\LearningPlanProfileType;
 use App\Models\MyLearningPlanFile;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -35,12 +36,20 @@ class LearningPlanController extends Controller
         }
 
         $t = new MyLearningPlan();
-        $t->profile_type_id = $request->profile_type;
         $t->title = $request->title;
         $t->description = $request->description;
         $t->image = $filename;
         $t->save();
 
+        if ($request->profile_type) {
+            $pt = json_decode($request->profile_type);
+            foreach($pt as $p){
+                $lppt = new LearningPlanProfileType();
+                $lppt->profile_type_id = $p->id;
+                $lppt->learning_plan_id = $t->id;
+                $lppt->save();
+            }
+        }
         return response(["status" => "success", "res" => $t], 200);
         }
     }
@@ -66,15 +75,26 @@ class LearningPlanController extends Controller
                 $destinationPath = public_path() . '/learning-plan-files';
                 $uploadedFile->move($destinationPath, $filename);
                 if ($t->image) {
-                    unlink($destinationPath . '/' . $t->image);
+                    if(file_exists($destinationPath . '/' . $t->image)){
+                        unlink($destinationPath . '/' . $t->image);
+                    }
                 }
                 $t->image = $filename;
             }
         }
         $t->title = $request->title;
         $t->description = $request->description;
-        $t->profile_type_id = $request->profile_type;
         $t->save();
+        if ($request->profile_type) {
+            $pt = json_decode($request->profile_type);
+            LearningPlanProfileType::where('learning_plan_id', $t->id)->delete();
+            foreach($pt as $p){
+                $lppt = new LearningPlanProfileType();
+                $lppt->profile_type_id = $p->id;
+                $lppt->learning_plan_id = $t->id;
+                $lppt->save();
+            }
+        }
         return response(["status" => "success", "res" => $t], 200);
     }
 
@@ -85,9 +105,14 @@ class LearningPlanController extends Controller
     public function get_learning_plan($id)
     {
         $ca = MyLearningPlan::with('files')
-            ->join("profile_types", 'profile_types.id', 'my_learning_plans.profile_type_id')
-            ->select('my_learning_plans.*', 'profile_types.profile_type')
+
             ->where('my_learning_plans.id', $id)->first();
+
+            $ca->profile_type = LearningPlanProfileType::join('profile_types', 'profile_types.id', 'learning_plan_profile_types.profile_type_id')
+            ->select('profile_types.id', 'profile_types.profile_type as name')
+            ->where('learning_plan_id', $id)
+            ->get();
+
         $path = url('/public/learning-plan-files/');
         $ca->image = $path . '/' . $ca->image;
         return response(["status" => "success", "res" => $ca, 'path' => $path], 200);
@@ -102,12 +127,14 @@ class LearningPlanController extends Controller
         $path = url('/public/learning-plan-files/');
         $user = Auth::guard('api')->user();
         $company = CompanyEmployee::where('user_id', $user->id)->first();
-        $ca = MyLearningPlan::with('files');
-        $ca = $ca->join("profile_types", 'profile_types.id', 'my_learning_plans.profile_type_id')
-            ->select('my_learning_plans.*', 'profile_types.profile_type');
-        if($company){
-            $ca = $ca->where('profile_types.id',$company->profile_type_id);
-        }
+        $ca = MyLearningPlan::with(['files','profileType' => function ($q) {
+            $q->join('profile_types', 'profile_types.id', 'learning_plan_profile_types.profile_type_id')->pluck('profile_types.profile_type');
+        }])->select('*');
+        // $ca = $ca->join("profile_types", 'profile_types.id', 'my_learning_plans.profile_type_id')
+        //     ->select('my_learning_plans.*', 'profile_types.profile_type');
+        // if($company){
+        //     $ca = $ca->where('profile_types.id',$company->profile_type_id);
+        // }
             $ca = $ca->paginate(10);
         return response(["status" => "success", "res" => $ca, "path" => $path], 200);
     }
@@ -121,12 +148,18 @@ class LearningPlanController extends Controller
         $path = url('/public/learning-plan-files/');
         $user = Auth::guard('api')->user();
         $company = CompanyEmployee::where('user_id', $user->id)->first();
-        $ca = MyLearningPlan::with('files');
-        $ca = $ca->join("profile_types", 'profile_types.id', 'my_learning_plans.profile_type_id')
-            ->select('my_learning_plans.*', 'profile_types.profile_type')
+        
+        // $ca = MyLearningPlan::with('files');
+        // $ca = $ca->join("profile_types", 'profile_types.id', 'my_learning_plans.profile_type_id')
+        //     ->select('my_learning_plans.*', 'profile_types.profile_type')
+        //     ->where('profile_types.id',$company->profile_type_id)
+        //     ->limit(6)
+        //     ->get();
+        $ca = MyLearningPlan::with(['files','profileType' => function ($q)use($company) {
+            $q->join('profile_types', 'profile_types.id', 'learning_plan_profile_types.profile_type_id')
             ->where('profile_types.id',$company->profile_type_id)
-            ->limit(6)
-            ->get();
+            ->pluck('profile_types.profile_type');
+        }])->select('*')->limit(6)->get();
         return response(["status" => "success", "res" => $ca, "path" => $path], 200);
     }
